#2029 Move document selection to its own component
